requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
eth-account==0.10.0
web3==6.15.1
solders==0.19.0
bip-utils==2.9.1
requests==2.31.0
aiohttp==3.9.1
base58==2.1.1
